LayoutTests/platform/ios-sim-deprecated/storage/indexeddb/dont-wedge-expected.txt

CONSOLE MESSAGE: line 22: TypeError: undefined is not an object (evaluating 'request.onerror = unexpectedBlockedCallback')
Check that a page reloaded during an in-flight upgradeneeded event doesn't hang.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;

dbname = "dont-wedge.html"
dbname1 = dbname + '1'
dbname2 = dbname + '2'

deleteDatabase1():
indexedDB.deleteDatabase(dbname1)
FAIL indexedDB.deleteDatabase(dbname1) threw exception TypeError: null is not an object (evaluating 'indexedDB.deleteDatabase')
FAIL Unexpected error: TypeError: undefined is not an object (evaluating 'request.onerror = unexpectedBlockedCallback')
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
